Simplify.

+++ src/main/java/org/rumbledb/runtime/functions/input/UrlValidator.java
@@ -2,7 +2,6 @@ package org.rumbledb.runtime.functions.input;
 
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FileContext;
-import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.UnsupportedFileSystemException;
 import org.rumbledb.exceptions.CannotRetrieveResourceException;
@@ -18,7 +17,7 @@ public class UrlValidator {
 
     private static final String[] allowedSchemes = { "file", "hdfs", "s3", "s3a", "s3n", "wasb", "gs", "root" };
 
-    public static boolean exists(String url, ExceptionMetadata metadata) {
+    public static URI validateURI(String url, ExceptionMetadata metadata) {
         if (url.isEmpty()) {
             throw new CannotRetrieveResourceException(
                     "No path provided!",
@@ -34,6 +33,11 @@ public class UrlValidator {
         if (locator.isAbsolute() && !Arrays.asList(allowedSchemes).contains(locator.getScheme())) {
             throw new OurBadException("Cannot interpret URL: " + url);
         }
+        return locator;
+    }
+
+    public static boolean exists(String url, ExceptionMetadata metadata) {
+        URI locator = validateURI(url, metadata);
         {
             try {
                 FileContext fileContext = FileContext.getFileContext();
@@ -55,47 +59,20 @@ public class UrlValidator {
         }
     }
 
-    public static boolean isDirectory(String url, ExceptionMetadata metadata) {
-        if (url.contains("*")) {
-            throw new CannotRetrieveResourceException(
-                    "Path cannot contain *!",
-                    metadata
-            );
-        }
-        URI locator = null;
-        try {
-            locator = new URI(url);
-            FileContext fileContext = FileContext.getFileContext();
-            Path path = new Path(url);
-            FileStatus status = fileContext.getFileStatus(path);
-            return status.isDirectory();
-        } catch (IOException e) {
-            e.printStackTrace();
-            throw new CannotRetrieveResourceException(
-                    "Error while accessing the " + locator.getScheme() + " filesystem.",
-                    metadata
-            );
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new OurBadException(
-                    "An unexpected exception happened while testing for directory.",
-                    metadata
-            );
-        }
-    }
-
     public static boolean delete(String url, ExceptionMetadata metadata) {
+        URI locator = validateURI(url, metadata);
         if (url.contains("*")) {
             throw new CannotRetrieveResourceException(
                     "Path cannot contain *!",
                     metadata
             );
         }
-        URI locator = null;
         try {
-            locator = new URI(url);
             FileContext fileContext = FileContext.getFileContext();
             Path path = new Path(url);
+            if (!fileContext.util().exists(path)) {
+                throw new OurBadException("Cannot delete file that does not exist: " + url);
+            }
             return fileContext.delete(path, true);
         } catch (IOException e) {
             e.printStackTrace();
@@ -113,17 +90,20 @@ public class UrlValidator {
     }
 
     public static FSDataInputStream getDataInputStream(String url, ExceptionMetadata metadata) {
+        URI locator = validateURI(url, metadata);
         if (url.contains("*")) {
             throw new CannotRetrieveResourceException(
                     "Path cannot contain *!",
                     metadata
             );
         }
-        URI locator = null;
         try {
             locator = new URI(url);
             FileContext fileContext = FileContext.getFileContext();
             Path path = new Path(url);
+            if (!fileContext.util().exists(path)) {
+                throw new OurBadException("Cannot read file that does not exist: " + url);
+            }
             return fileContext.open(path);
         } catch (IOException e) {
             e.printStackTrace();
